requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-cleanup
django-cors-headers
django-filter
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
drf-nested-routers
drf-yasg
idna
inflection
Markdown
oauthlib
packaging
Pillow
pycparser
PyJWT
python3-openid
pytz
PyYAML
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
uritemplate
urllib3
gunicorn
